+++ examples/demo.php
@@ -6,11 +6,8 @@ require_once dirname(__DIR__) . '/vendor/autoload.php';
 use Jieba\Jieba;
 use Jieba\Options;
 use Jieba\Option\Dict;
-use Jieba\Option\Mode;
 
-$jieba = new Jieba(
-    (new Options())->setDict(new Dict(Dict::BIG))->setMode(new Mode(Mode::TEST))
-);
+$jieba = new Jieba((new Options())->setDict(new Dict(Dict::BIG)));
 
 $seg_list = $jieba->cut("怜香惜玉也得要看对象啊！");
 var_dump($seg_list);