+++ examples/demo_extract_tags.php
@@ -7,11 +7,8 @@ use Jieba\Jieba;
 use Jieba\JiebaAnalyse;
 use Jieba\Options;
 use Jieba\Option\Dict;
-use Jieba\Option\Mode;
 
-$jieba = new Jieba(
-    (new Options())->setDict(new Dict(Dict::BIG))->setMode(new Mode(Mode::TEST))
-);
+$jieba = new Jieba((new Options())->setDict(new Dict(Dict::BIG)));
 $tags = JiebaAnalyse::singleton()->extractTags(
     $jieba->cut(file_get_contents(dirname(__DIR__) . '/tests/dict/lyric.txt')),
     10