+++ examples/demo_user_dict.php
@@ -6,11 +6,8 @@ require_once dirname(__DIR__) . '/vendor/autoload.php';
 use Jieba\Jieba;
 use Jieba\Options;
 use Jieba\Option\Dict;
-use Jieba\Option\Mode;
 
-$jieba = new Jieba(
-    (new Options())->setDict(new Dict(Dict::SMALL))->setMode(new Mode(Mode::TEST))
-);
+$jieba = new Jieba((new Options())->setDict(new Dict(Dict::SMALL)));
 $seg_list = $jieba->cut("李小福是创新办主任也是云计算方面的专家");
 var_dump($seg_list);