+++ src/Jieba/Options.php
@@ -3,7 +3,6 @@
 namespace Jieba;
 
 use Jieba\Option\Dict;
-use Jieba\Option\Mode;
 
 /**
  * Class Options
@@ -13,17 +12,11 @@ use Jieba\Option\Mode;
 class Options
 {
     const DICT = 'dict';
-    const MODE = 'mode';
 
     /**
      * @var Dict
      */
     protected $dict;
-
-    /**
-     * @var Mode
-     */
-    protected $mode;
 
     /**
      * Options constructor.
@@ -32,9 +25,7 @@ class Options
      */
     public function __construct(array $options = [])
     {
-        $this
-            ->setDict(!empty($options[self::DICT]) ? new Dict($options[self::DICT]) : new Dict())
-            ->setMode(!empty($options[self::MODE]) ? new Mode($options[self::MODE]) : new Mode());
+        $this->setDict(!empty($options[self::DICT]) ? new Dict($options[self::DICT]) : new Dict());
     }
 
     /**
@@ -48,9 +39,6 @@ class Options
 
         if (!$this->isDictionaryValid($options[self::DICT])) {
             throw new Exception("invalid dictionary '{$options[self::DICT]}' specified");
-        }
-        if (!$this->isModeValid($options[self::MODE])) {
-            throw new Exception("invalid mode '{$options[self::MODE]}' specified");
         }
 
         return $options;
@@ -71,25 +59,6 @@ class Options
     public function setDict(Dict $dict): Options
     {
         $this->dict = $dict;
-
-        return $this;
-    }
-
-    /**
-     * @return Mode
-     */
-    public function getMode(): Mode
-    {
-        return $this->mode;
-    }
-
-    /**
-     * @param Mode $mode
-     * @return $this
-     */
-    public function setMode(Mode $mode): Options
-    {
-        $this->mode = $mode;
 
         return $this;
     }